// src/api/model/jobInfo.ts
/**
 * Generated by orval v7.0.1 🍺
 * Do not edit manually.
 * 5-thirty
 * backend data aggregate for 6 o'clock
 * OpenAPI spec version: 0.0.3
 */

export interface JobInfo {
  cancelled: boolean;
  endedAt?: Date;
  failed: boolean;
  id: number;
  key?: string;
  startedAt?: Date;
  succeeded: boolean;
  title: string;
}